chore(plugin): Split css.ts into two files: media.ts and plugin.ts, so unnecessary code won't be shipped in JS bundle
